requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
bidict==0.21.4
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
dnspython==2.1.0
elastic-transport==8.4.0
elasticsearch==8.5.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.2.0
lxml==4.6.4
MarkupSafe==2.0.1
nltk==3.6.5
outcome==1.1.0
pdfminer.six==20211012
pycparser==2.21
pymongo==4.0
pyOpenSSL==21.0.0
python-dotenv==0.19.2
python-engineio==4.3.0
python-socketio==5.5.0
regex==2021.11.10
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tqdm==4.64.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
